Make the tests build with a current pspdev toolchain
Most test directories no longer compiled under pspdev's GCC 15, which meant
gentest.py aborted before reaching the PSP for any test that shared a
directory with a broken one. All 100-odd directories build clean now.

Nothing here regenerates a .prx - the committed binaries are untouched.

The toolchain-driven changes:

- C23 makes implicit function declarations an error, and the SDK headers no
  longer pull each other in transitively. Added the includes that were being
  got for free, and real prototypes for the handful of functions no pspsdk
  header declares (the mutex calls, sceKernelPower*, sceKernelGetModuleId).
- GCC 10's -fno-common turned tentative definitions shared between a test and
  the shared.o linked beside it into duplicate symbols; those globals are now
  file-local. Same story for schedf, which a dozen tests deliberately override
  - the copy in libcommon is weak now, which is what the old linker did in
  effect.
- C99 inline semantics emit no out-of-line definition, so the bare `inline`
  helpers in the shared test headers are `static inline`. In tests/libs/xlib
  they're declared non-static in headers, so there they're plain functions.
- Renames the SDK made: SceIoStat's sce_st_ctime and friends, pspTime ->
  ScePspDateTime with singular field names, sendCommandi -> sceGuSendCommandi,
  sceJpeg_67F0ED84 -> sceJpegCsc. Dropped the local declarations of types and
  functions the SDK now provides, and gave the savedata test's own structs the
  "2" suffix it already used for SceUtilitySavedataParam2 so they stop
  colliding.
- Kept -Wint-conversion, -Wincompatible-pointer-types and -Wnarrowing as
  warnings. Handing the kernel a NULL where a SceUID goes, or 0xAB as a char,
  is the point of those tests, and casting would delete the thing being tested.
- Binutils won't assemble "jalr $t0, $t0" - the rd == rs case cpu_branch
  exists to test - so that one is encoded by hand. vector.c and vregs.cpp
  stored VFPU registers through an "m" constraint on a static array, which now
  expands to %lo(sym)($reg); sv.q can't encode that, so they take the address
  in a register.
- libpspsnd is gone, sceMp3LowLevel* and SceMp3InitArg changed shape, and
  libpspkernel now collides with newlib over strtol - so iodrv and loadexec
  import the two or three kernel functions they need directly, the way other
  tests here already do.

Two Makefiles were quietly broken rather than stale: tests/cpu/cpu_alu had its
TARGETS overwritten in 7e4a1e7 with a file belonging to tests/cpu/crash, and
tests/ctrl and tests/video/psmfplayer pointed at imports objects in their own
directory instead of the one next door. gpu/commands/light.cpp referenced a
GU_UNKNOWN_LIGHT_COMPONENT that never existed; sceGuLight maps components == 8
to GE light component 2, which is the "A + pow(D)" case it wants.

Verified on hardware (6.61) that rebuilt binaries still match .expected for
cpu_branch, vector, vregs, lwmutex/lock, semaphores/create, sysmem/partition,
mp3/getbitrate and mp3/getframenum. Note that rebuilding is not free: time_t
is 64-bit in the current toolchain, so rtc/convert's sceRtcSetTime_t cases
marshal differently than the committed .prx does, and the PRXs come out ~35%
larger. Another reason to leave the committed binaries alone.

+++ tests/font/newlib.cpp
@@ -48,7 +48,7 @@ void testNewLib(const char *title, FontNewLibParams *params, bool useError = tru
 	uint error = -1;
 	g_allocated = 0;
 	FontLibraryHandle libID = sceFontNewLib(params, useError ? &error : NULL);
-	if (libID <= 0) {
+	if ((int)libID <= 0) {
 		checkpoint("%s: Failed (%08x, %08x)", title, libID, error);
 	} else if (error != 0) {
 		checkpoint("%s: Error (%08x, %08x)", title, libID, error);
